backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
PyJWT
python-decouple
python-multipart
asyncpg
